\name{generalized.triangular.numbers-package}
\alias{generalized.triangular.numbers-package}
\alias{generalized.triangular.numbers}
\docType{package}
\title{
  Compute the generalized triangular numbers
}
\description{
See \code{\link{gtn}}
}
\author{
Maintainer: Chase Benoist <csbcodev@gmail.com>
}
\keyword{ package }
